checkpoint switching to npm and webpack with many upgrades
In the react ecosystem, npm for package management and webpack/friends are
the way things are done.  I'd started out using RequireJS/volo/AMD/alameda
for consistency benefits with gaia mail, but the costs got to be too much
as I latched react versions and found it difficult to reuse existing
libraries.

It's my hope that this change makes it easier to stay up-to-date, reuse
libraries, and make it easier for others to contribute and understand what's
going on.  I believe I've done things in such a way that gaia mail can
continue to use the GELAM backend using its existing RequireJS/volo/alameda
implementation thanks to webpack's understanding of AMD.  I also believe
webpack's chunking support is sufficient to give us many of the benefits
r.js's layers mechanism provides for gaia mail.  Do note that I've restrained
the chunks for now because the worker interaction is sufficiently complex that
I don't want to entirely deal with it right now.