checkpoint conversation list working, LiveItemMixin issues, changes planned
We've been using two mechanisms to trigger update of React components based on
events from GELAM:

1. EntireList and WindowedList pass the `serial` of the view and the items into
   their ReactList and the widget items as props.  This works with
   PureRenderMixin, but for evolutionary reasons we ended up with
   slice_item_mixin.js that can be replaced by PureRenderMixin or equivalent.

2. The header items use live_item_mixin that attempts to do a similar thing
   inside a component itself by adding a `change` listener and updating the
   state whenever anything changes.

Since mix-ins are only viable using React.createClass() and that's
idiomatically on the way out, our options are basically:

- higher-order-component like react-redux's connect() or react-intl's
  (connect-inspired) injectIntl().  These basically interpose an
  additional widget that declares a `contextTypes` entry and propagates all
  of the props they receives plus the extras they add from the context.

- Have our "viewing" reducer expose the serial numbers/similar values as
  explicit state inside the redux store that works with PureRenderMixin-style
  logic.  So `conversation` might also have a sibling `conversationSerial`
  and `conversationsView` might also have `conversationsViewMetaSerial` for
  consumers that care about the tocMeta for the view.  This does also require
  that the reducer registers for events and dispatches actions against itself.

- Keep doing the mix-in type thing with state, but somehow with a decorator.

- Have the components explicitly do what the mix-in does inside of themselves.

Right now, I am heavily leaning towards having the reducers expose the serial
numbers into the redux state.  The use of sibling values seems a little hacky,
but tupling the view with its serial in a new encapsulating object to effect
the change from an === perspective doesn't seem any cleaner.  Additionally, the
use of sibling values does allow the inherent view serial to be decoupled from
the tocMeta serial that we can create.  This matters in the case of headers
versus lists where the headers probably only care about the tocMeta and the
lists only care about the items.

One nice positive is that we get to work within the redux connect() idiom and
avoid having to add any other interposing widgets.  Plus the serial change is
made visible in the state which hopefully is more intuitive than mix-ins
maintaining their own state machines inside the components.  A downside to this
is that anyone wanting to play with redux devtools may be very disappointed
that we generate an incredible number of internal dispatches that exist only to
trigger updates and exist on top of underlying mutable state.  So if they try
and do any time travelling, they'll find that most travelling results in no
visible changes unless the state changes were actually due to user-initiated
interaction.  Perhaps there's some way we can tag automatic actions versus
user-initiated to assist such devtools?