redux view flow and 3-pane working with scrolling after giving up on flexbox
The hackish integration of redux and gelam's view slice and per-item serial
tracking now works reasonably happily.

Layout-wise:

It's clear I need a freshly awake brain for this, but in the interest of
saving awake-me some trouble, or at least giving awake-me something to
laugh at, here are various things I tried:

The eternal desire to enforce xul-style border-layout-esque things where
we don't hardcode the sizes of the things on the periphery once again
loses out.  I sleepily tried various hacks to interpose "position: relative"
boxes so that react-split-pane's use of "height: 100%" would not net out to
a height greater than 100%, but that tended to either take the box out of
the flow or run into the problem where the constraints were not being
propagated.

I also tried hacking react-split-pane to reduce its dependencies on using
100% but that tended to not cause the area we want to scroll to become
scrolling and instead grew the whole document.  This all very much suggests
I need to refresh my understanding of the flexbox layout model.

Some other npm packages were briefly investigated, but none seemed like a clear
win.  The most promising seemed to be trying to do more of the layout logic
itself.  And while in frustration I could see agreeing with that, it does seem
preferable to instead come back up to speed on flexbox.