Beginning of yak shaving upgrades to get webpack build working.
Webpack's dynamic builds seemed okay-ish but was going off the rails attempting
to process non-JS gelam submodules.  I started the process of updating things,
but pulling that thread is going to result in a bunch of updates and code
transforms that weren't strictly necessary but should allow build
simplifications and consistency with other related projects.