Import my forked version of react-list for the async quantized list.
Copying over react-list.es6 from
https://github.com/asutherland/react-list/tree/quantized-variable-list
renamed with the intent to refactor it to work with modern react and
only be used for the windowed list case.  Existing usages of react-list
are intended to be removed unless there's a demonstrated need to keep
things out of the DOM for performance reasons.  (Better to err on
the side of supporting ctrl-f.)

It's probably appropriate to revisit the state of the art for virtual
lists, although the specific worker data model in use here and
intentionally stable item heights would seem to remain unique based on
a quick survey.  (And from general web browsing experience, it seems
that the dynamic sizing virtual lists out there have far from solved
the problem.)