get rid of webpack-dev-server and all its dependencies
There may be some detritus in terms of make targets, etc., but the reality is
I'm not using the live reloading at this time and it's not clear that live
reloading is likely to be safe for the usage.  Also... in general we'd want
routes/history API to make it so that reloading the page shouldn't actually
lose any state, so that's a good dogfooding goal.