Get phabricator "patch-paths" virtual folder root working
With this change, patches under review are fetched and reviews are placed in
virtual folders according to the directories they touch.

This involved creating a new GELAM task decorator mechanism to enable the
on-demand folder creation.

This also involved vendoring/forking the parse-git-patch module because it
would fail to parse git patches that didn't include all of the email-style
metadata (From, Subject, Date).